requirements.txt
asgiref==3.2.10
beautifulsoup4==4.9.1
Django==3.1
django-bootstrap4==2.2.0
django-filter==2.3.0
django-redis==4.12.1
djangorestframework==3.11.1
importlib-metadata==1.7.0
Markdown==3.2.2
pytz==2020.1
redis==3.5.3
soupsieve==2.0.1
sqlparse==0.3.1
zipp==3.1.0
